split fixtures
